Move creatable file types to /models
